Refactor
Rename `BuildCache` as `BuildFileCache`.

Rename field of `BuildFileCache` as `fileCache` (`times` was a misnomer).

Move `FileCacheInfo` from `Stack.Types.Package` to be co-located with `BuildFileCache`.

Use type synonym `FileCache`.

Move `toCachePkgSrc` to `Stack.Build.ConstructPlan` (the only place it is used).

Move functions from `Stack.Types.ConfigureOpts` to `Stack.ConfigureOpts`, to avoid cyclic module dependencies.

Rename `configureOpts` to `configureOptsFromBase` to avoid clash with `ConfigCache` field name.
