+++ .stan.toml
@@ -1,11 +1,11 @@
 # Partial: base/head
 # Usage of partial function 'head' for lists
 [[ignore]]
-  id = "OBS-STAN-0001-1XF2BN-54:1"
+  id = "OBS-STAN-0001-1XF2BN-73:1"
 # ✦ Category:      #Partial #List
 # ✦ File:          src\Stack\Storage\Project.hs
 #
-#   54 ┃ share [ mkPersist sqlSettings
+#   73 ┃ share [ mkPersist sqlSettings
 
 # Partial: base/head
 # Usage of partial function 'head' for lists
@@ -151,14 +151,14 @@
 
 # Anti-pattern: Data.ByteString.Char8.pack
 [[ignore]]
-  id = "OBS-STAN-0203-tuE+RG-248:24"
+  id = "OBS-STAN-0203-tuE+RG-249:24"
 # ✦ Description:   Usage of 'pack' function that doesn't handle Unicode characters
 # ✦ Category:      #AntiPattern
 # ✦ File:          src\Stack\Build\ExecutePackage.hs
 #
-#   247 ┃
-#   248 ┃   newConfigFileRoot <- S8.pack . toFilePath <$> view configFileRootL
-#   249 ┃                        ^^^^^^^
+#   248 ┃
+#   249 ┃   newConfigFileRoot <- S8.pack . toFilePath <$> view configFileRootL
+#   250 ┃                        ^^^^^^^
 
 # Anti-pattern: Data.ByteString.Char8.pack
 [[ignore]]
