+++ package.yaml
@@ -201,6 +201,7 @@ library:
   - Stack.Config.ConfigureScript
   - Stack.Config.Docker
   - Stack.Config.Nix
+  - Stack.ConfigureOpts
   - Stack.ConfigCmd
   - Stack.Constants
   - Stack.Constants.Config
